// astr-ai-code-mother-frontend/lib/api/generated/model/appVO.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { UserVO } from './userVO';

export interface AppVO {
  id?: number;
  appName?: string;
  cover?: string;
  initPrompt?: string;
  codeGenType?: string;
  deployKey?: string;
  deployedTime?: string;
  priority?: number;
  userId?: number;
  createTime?: string;
  updateTime?: string;
  user?: UserVO;
}